// packages/api-server/src/authorization/enums/Permission.ts
export enum Permission {
  ADMIN = 'ADMIN',
  ACCOUNT_CREATE = 'ACCOUNT_CREATE',
  ACCOUNT_READ = 'ACCOUNT_READ',
  ACCOUNT_UPDATE = 'ACCOUNT_UPDATE',
  ACCOUNT_DELETE = 'ACCOUNT_DELETE',
  ACCOUNT_READ_OWNER = 'ACCOUNT_READ_OWNER',
  ACCOUNT_UPDATE_OWNER = 'ACCOUNT_UPDATE_OWNER',
  ACCOUNT_DELETE_OWNER = 'ACCOUNT_DELETE_OWNER',
}